[feat/#71] 멀티 뷰타입 및 무한 스크롤 설정
