+++ domain/src/main/java/com/depromeet/domain/entity/request/home/MySeatRecordRequest.kt
@@ -2,7 +2,7 @@ package com.depromeet.domain.entity.request.home
 
 data class MySeatRecordRequest(
     val page: Int ?= null,
-    val size: Int ?= 50,
+    val size: Int ?= 10,
     val year: Int ?= 2024,
     val month: Int ?= null,
 )
